requirements.txt
ccxt==4.3.82
python-dotenv==1.0.1
requests==2.32.3
PyYAML==6.0.2
pytz==2024.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
ta==0.11.0